Add testing for additional reducers and actions:
