+++ actions/actions.test.js
@@ -61,5 +61,30 @@ describe('actions', () => {
 
         const result = actions.toggleCurrentRide(num)
         expect(result).toEqual(expectedAction)
+    });
+
+    it('should have a type of TOGGLE_LOADING', () => {
+        const bool = !bool;
+
+        const expectedAction = {
+            type: 'TOGGLE_LOADING',
+            bool
+        };
+
+        const result = actions.toggleLoading(bool);
+        expect(result).toEqual(expectedAction);
+
+    });
+
+    it('should have a type of TOGGLE_RIDE_ATTENDANCE', () => {
+        const bool = !bool;
+
+        const expectedAction = {
+            type: 'TOGGLE_RIDE_ATTENDANCE',
+            bool
+        };
+
+        const result = actions.toggleRideAttendance(bool)
+        expect(result).toEqual(expectedAction)
     })
 })
